Math stuff

+++ src/main/java/MyMathCalc.java
@@ -5,7 +5,7 @@ public class MyMathCalc {
     public int multiply (int a, int b) {
         return a * b;
     }
-    public int divide (int a, int b) {
+    public double divide (int a, int b) {
         return a / b;
     }
 }
